Move InstalledSnaps helper to snapstate.
